requirements.txt
astroid==1.5.3
certifi==2017.7.27.1
chardet==3.0.4
click==6.7
docopt==0.6.2
elasticsearch==5.4.0
elasticsearch-dsl==5.3.0
Flask==0.12.2
idna==2.6
isort==4.2.15
itsdangerous==0.24
Jinja2==2.9.6
lazy-object-proxy==1.3.1
MarkupSafe==1.0
mccabe==0.6.1
pylint==1.7.4
python-dateutil==2.6.1
requests==2.18.4
six==1.11.0
urllib3==1.22
Werkzeug==0.12.2
wrapt==1.10.11
